Moved projections to core submodule
