+++ core/src/main/java/uk/ac/ebi/eva/commons/core/models/mongodb/projections/SimplifiedAnnotation.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.commons.mongodb.projections;
+package uk.ac.ebi.eva.commons.core.models.mongodb.projections;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Field;
